worldsim_data_fetcher_full/requirements.txt
requests>=2.32.0
tqdm>=4.66.0
pyyaml>=6.0.1
python-dateutil>=2.9.0.post0
pyshp>=3.0.3
